Move block context unmarshalling logic
